requirements.txt
diff-match-patch==20121119
Django==2.1
django-ckeditor==5.6.1
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django-import-export==1.0.1
django-js-asset==1.1.0
django-recaptcha==1.4.0
django-sendgrid-v5==0.7.0
et-xmlfile==1.0.1
future==0.17.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.6
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-http-client==3.1.0
pytz==2018.5
PyYAML==3.13
sendgrid==5.6.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
